Pull out mappingsFromMap into /lib folder
